simplifies the logic to request user id
