+++ src/firebase/firebase.utils.js
@@ -1,26 +1,18 @@
 import firebase from "firebase/app";
 import "firebase/firebase-firestore";
 import "firebase/firebase-auth";
-
-const firebaseConfig = {
-  apiKey: "AIzaSyA2MPd6if-SORKEnBlbOO1rpQHxyD4pC4k",
-  authDomain: "trivia-quiz-35a66.firebaseapp.com",
-  projectId: "trivia-quiz-35a66",
-  storageBucket: "trivia-quiz-35a66.appspot.com",
-  messagingSenderId: "1060181821141",
-  appId: "1:1060181821141:web:57d3089d7972b3d4ef445c",
-  measurementId: "G-Y1JV0SLE4T",
-};
+import firebaseConfig from "./firebase_config";
 
 firebase.initializeApp(firebaseConfig);
 
 const auth = firebase.auth();
 const fireStore = firebase.firestore();
 
-const provider = new firebase.auth.GoogleAuthProvider();
-provider.setCustomParameters({ prompt: "select_account" });
-
-const signInWithGoogle = () => auth.signInWithPopup(provider);
+const signInWithGoogle = () => {
+  const provider = new firebase.auth.GoogleAuthProvider();
+  provider.setCustomParameters({ prompt: "select_account" });
+  auth.signInWithPopup(provider);
+};
 
 const createUserProfileDocument = async (userAuth) => {
   if (!userAuth) {
@@ -44,17 +36,13 @@ const createUserProfileDocument = async (userAuth) => {
   return userReference;
 };
 
-const setUserScore = async ({
-  currentUserId,
-  category,
-  score,
-  totalQuestions,
-}) => {
-  if (!currentUserId) {
-    return;
-  }
+const setUserScore = async ({ category, score, totalQuestions }) => {
+  const currentUser = firebase.auth().currentUser;
+  if (!currentUser) return;
 
-  const userReference = fireStore.doc(`/users/${currentUserId}`);
+  let userId = currentUser.uid;
+
+  const userReference = fireStore.doc(`/users/${userId}`);
   const data = {
     category: category,
     score: `${score}/${totalQuestions}`,
@@ -68,10 +56,12 @@ const setUserScore = async ({
   }
 };
 
-const getUserScores = async (userId) => {
-  if (!userId) {
-    return;
-  }
+const getUserScores = async () => {
+  const currentUser = firebase.auth().currentUser;
+  if (!currentUser) return;
+
+  console.log("here");
+  let userId = currentUser.uid;
 
   const userReference = fireStore.doc(`/users/${userId}`);
   const user = await userReference.get();
