+++ src/firebase/firebase_config.js
@@ -1,0 +1,11 @@
+const firebaseConfig = {
+  apiKey: "AIzaSyA2MPd6if-SORKEnBlbOO1rpQHxyD4pC4k",
+  authDomain: "trivia-quiz-35a66.firebaseapp.com",
+  projectId: "trivia-quiz-35a66",
+  storageBucket: "trivia-quiz-35a66.appspot.com",
+  messagingSenderId: "1060181821141",
+  appId: "1:1060181821141:web:57d3089d7972b3d4ef445c",
+  measurementId: "G-Y1JV0SLE4T",
+};
+
+export default firebaseConfig;
